Update SegmentHippoBoxedCenter.py
Only uses a single positive point to detect the CV (SAM accuracy is pretty much the same with 1 point vs 2 points that are right next to each other).

Instead of using random positions for negative points, it steps through the image to grab points at equal proportions of every image.